+++ source/luascript.h
@@ -35,7 +35,7 @@ extern "C" struct lua_State;
 #ifdef YUR_HIGH_LEVELS
 typedef int64_t exp_t;
 #else
-typedef unsigned long exp_t;
+typedef uint32_t exp_t;
 #endif //YUR_HIGH_LEVELS
 
 class LuaScript
